requirements.txt
Django==2.2.5
requests==2.22.0
djangorestframework==3.10.2
freezegun==0.3.12
django-optimistic-lock==1.0.0
markdown==3.1.1
